Update OrderPage.html
Updated the images of the drinks. Also changed "Ready for Payment." to "Payment Successful." on the Pay For Order button click. Made the starbucks image inline with the Cashier's App heading. Also added padding to the bottoms of the drinks for easier reading. Drink names are bolded now. Changed "finish Order" button to "Pay For Order" to simulate the customer paying.